Done with bug fixing for now, refactored codebase, and finished the feature, will move to do pull request
